Major API changes
timer - adding new timer API.  The old API was unimplemented.  The new
API should work across several devices and provide a consistent way to
set up and use timers.
intmath and gpio - many of these functions have been inlined and added
to the header files so they can be used in ISRs.
ui - These functions aren't used in the modern PowerPallet code and are
deprecated in favor of the newer UI library.
vnh_bridge - This is being modified to use the new timer API directly.

This code is not yet fully functional.